--- docs/reference/source/nashpy.fictitious_learning.rst
@@ -1,13 +1,0 @@
-nash\.learning package
-=================================
-
-Submodules
-----------
-
-nashpy\.learning\.\fictitious\_learning\.get\_best\_response\_to\_play\_count module
-------------------------------------------------------------------------------------
-
-.. automodule:: nashpy.learning.fictitious_learning
-    :members:
-    :undoc-members:
-    :show-inheritance:
