+++ docs/reference/source/nashpy.learning.rst
@@ -1,0 +1,13 @@
+nash\.learning package
+======================
+
+Submodules
+----------
+
+nashpy\.learning\.\fictitious\_play module
+------------------------------------------
+
+.. automodule:: nashpy.learning.fictitious_play
+    :members:
+    :undoc-members:
+    :show-inheritance:
