+++ src/nashpy/learning/fictitious_play.py
